3lab/proj/Mirach_laba3/requirements.txt
fastapi
uvicorn
sqlalchemy
redislite
python-multipart
python-jose
pydantic
requests
celery
aiohttp
pydantic[email]
websockets